chore(dsp-api-error): refactor CatalogErrorToResponseTransformer to JsonObjectFromCatalogErrorTransformer
